Try to fix image upload with Multer

+++ controllers/postsController.js
@@ -2,6 +2,10 @@ const express = require("express");
 const { PrismaClient } = require("@prisma/client");
 const prisma = new PrismaClient();
 const generateSlug = require("../utilities/generateSlug");
+const fs = require("fs");
+const util = require("util");
+// const unlinkFile = util.promisify(fs.unlink);
+
 
 // errors handler
 const NotFound = require("../errors/NotFound");
@@ -75,36 +79,39 @@ async function create(req, res, next)
 	const validation = validationResult(req);
 	if (!validation.isEmpty())
 	{
-		return next(
-			new validationError("verificare i dati inseriti", validation.array())
-		);
+		return next(new validationError("Verificare i dati inseriti", validation.array()));
 	}
+	try
+	{
+		const { title, content, published, author, categoryId, tags } = req.body;
+		const imagePath = req.file ? req.file.path : null; // Controlla se c'è un file caricato
+		slug = await generateSlug(title); // Genera uno slug dal titolo
 
-	const { title, author, image, content, published } = req.body;
-	const slug = await generateSlug(title);
+		// Creazione del post
+		const cat = parseInt(categoryId);
+		const post = await prisma.post.create({
+			data: {
+				title,
+				slug,
+				content,
+				published: published === 'true', // Assicurati che published sia un booleano
+				author,
+				image: imagePath,
+				categoryId: cat, // Usa l'ID della categoria selezionata
+				// tags: {
+				// 	connect: tags.map(tag => ({ id: tag.id })) // Collega i tag esistenti
+				// }
+			}
+		});
 
-	const newPost = await prisma.post.create({
-		data: {
-			title,
-			slug,
-			author,
-			image,
-			content,
-			published,
-			categoryId,
-			userId,
-			tags: {
-				connect: req.body.tags.map(tag => ({ id: tag.id })),
-			},
-		},
-		include: {
-			category: true,
-			tags: true,
-		}
-	});
+		res.status(201).json({ message: "Post creato con successo", post });
+	} catch (error)
+	{
+		console.error(error);
+		res.status(500).send("Errore nella creazione del post");
+	}
+};
 
-	res.json(newPost);
-}
 
 
 // show - read a single post by slug
@@ -153,9 +160,7 @@ async function update(req, res, next)
 	let updateData = req.body;
 
 	const post = await prisma.post.findUnique({
-		where: {
-			slug: slug,
-		},
+		where: { slug },
 		include: {
 			category: true,
 			tags: true,
@@ -167,6 +172,23 @@ async function update(req, res, next)
 		return next(new NotFound(`Post not found with slug: ${slug}`));
 	}
 
+
+	// update image
+	If(updateData.image && post.image !== updateData.image);
+	{
+		if (post.image)
+		{
+			try
+			{
+				await unlinkAsync(post.image);
+			} catch (error)
+			{
+				console.log("Errore nella rimozione dell'immagine esistente:", error);
+			}
+		}
+	}
+
+	// update title and slug
 	if (req.body.title)
 	{
 		updateData.slug = await generateSlug(req.body.title);
@@ -197,7 +219,7 @@ async function update(req, res, next)
 	}
 
 	const updatedPost = await prisma.post.update({
-		where: { slug: slug },
+		where: { slug },
 		data: updateData,
 		include: {
 			category: true,
@@ -220,12 +242,31 @@ async function update(req, res, next)
 async function destroy(req, res, next)
 {
 	const { slug } = req.params;
+	const post = await prisma.post.findUnique({
+		where: { slug },
+	});
+
+	if (!post)
+	{
+		return next(new NotFound(`Post not found with slug: ${slug}`));
+	}
+
+	if (post.image)
+	{
+		try
+		{
+			await unlinkAsync(post.image);
+		} catch (error)
+		{
+			console.log("Errore nella rimozione dell'immagine esistente:", error);
+		}
+	}
 
 	await prisma.post.delete({
 		where: { slug }
 	});
 
-	return res.json({ message: "Post successfully deleted" });
+	res.json({ message: "Post successfully deleted" });
 }
